+++ src/main/java/br/com/fiap/bookorganizer/repositories/AutorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.bookorganizer.models.Autor;
 
-public interface AutorRepository extends JpaRepository<Autor, Integer> {
+public interface AutorRepository extends JpaRepository<Autor, Long> {
     
 }
